# openapi.yaml
openapi: 3.0.3
info:
  title: Warm Follow SaaS API
  version: 1.0.1
  description: |
    Single source of truth for frontend integration.
    All responses use snake_case. Auth uses JWT bearer tokens.
servers:
  - url: http://localhost:8080
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    WorkspaceId:
      name: X-Workspace-Id
      in: header
      required: false
      schema:
        type: string
        format: uuid
      description: Override current workspace context.
...
  responses:
...
    BadRequestResponse:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    UnauthorizedResponse:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    ForbiddenResponse:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    ConflictResponse:
      description: Conflict
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    ErrorResponse:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
          additionalProperties: true
        traceId:
          type: string
      example:
        code: LIMIT_CLIENTS_REACHED
        message: Client limit reached for current plan.
        details:
          limit: 200
          current: 200
        traceId: b3f1c2a8
    StatusResponse:
      type: object
      required: [status]
      properties:
        status:
          type: string
      example:
        status: ok
    AuthResponse:
      type: object
      required: [access_token, refresh_token, token_type, expires_in, user, workspace, role]
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        token_type:
          type: string
        expires_in:
          type: integer
          format: int64
        user:
          $ref: "#/components/schemas/UserResponse"
        workspace:
          $ref: "#/components/schemas/WorkspaceResponse"
        role:
          type: string
          enum: [owner, admin, member]
      example:
        access_token: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9
        refresh_token: 6f7c0b40-1f32-4f1b-8ab1-89b1b7cc0b91
        token_type: Bearer
        expires_in: 900
        user:
          id: 2b7c1dd6-8b8d-4b77-b2cb-9b0b3b97e7a0
          email: owner@example.com
          name: Owner User
          status: active
          last_login: "2026-01-20T10:20:00Z"
        workspace:
          id: 6d5c2e8a-4b0b-4da3-9a9f-3f8dd6d93e2a
          name: Acme Workspace
          plan: starter
          status: active
          billing_status: active
          created_at: "2026-01-20T10:10:00Z"
        role: owner
    UserResponse:
      type: object
      required: [id, email, name, status]
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        name:
          type: string
        status:
          type: string
          enum: [active, invited, disabled]
        last_login:
          type: string
          format: date-time
    WorkspaceResponse:
      type: object
      required: [id, name, plan, status]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        plan:
          type: string
          enum: [starter, pro, business]
        status:
          type: string
          enum: [trialing, active, past_due, suspended, canceled]
        billing_status:
          type: string
          enum: [trialing, active, past_due, canceled, unpaid, incomplete]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    WorkspaceContextResponse:
      type: object
      required: [workspace, role]
      properties:
        workspace:
          $ref: "#/components/schemas/WorkspaceResponse"
        role:
          type: string
          enum: [owner, admin, member]
    MeResponse:
      type: object
      required: [user, workspace, role, limits, usage, counts]
      properties:
        user:
          $ref: "#/components/schemas/UserResponse"
        workspace:
          $ref: "#/components/schemas/WorkspaceResponse"
        workspaces:
          type: array
          items:
            $ref: "#/components/schemas/WorkspaceContextResponse"
        role:
          type: string
          enum: [owner, admin, member]
        limits:
          $ref: "#/components/schemas/LimitsResponse"
        usage:
          $ref: "#/components/schemas/UsageResponse"
        counts:
          $ref: "#/components/schemas/CountsResponse"
      example:
        user:
          id: 2b7c1dd6-8b8d-4b77-b2cb-9b0b3b97e7a0
          email: owner@example.com
          name: Owner User
          status: active
          last_login: "2026-01-20T10:20:00Z"
        workspace:
          id: 6d5c2e8a-4b0b-4da3-9a9f-3f8dd6d93e2a
          name: Acme Workspace
          plan: starter
          status: active
          billing_status: active
          created_at: "2026-01-20T10:10:00Z"
        role: owner
        limits:
          plan: starter
          max_users: 3
          max_clients: 200
          retention_days: 365
        usage:
          period: "2026-01"
          emails_sent: 120
          sms_sent: 45
          sms_credits_balance: 10
          overage_cost_cents: 0
        counts:
          clients: 152
          users: 2
    CountsResponse:
      type: object
      required: [clients, users]
      properties:
        clients:
          type: integer
          format: int64
        users:
          type: integer
          format: int64
...
    DnsRecord:
      type: object
      required: [type, name, value]
      properties:
        type:
          type: string
          enum: [TXT, CNAME, MX]
        name:
          type: string
        value:
          type: string
        ttl:
          type: integer
        priority:
          type: integer
        purpose:
          type: string
          enum: [spf, dkim, dmarc]
    SendingDomainResponse:
      type: object
      required: [id, domain, status, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        domain:
          type: string
        status:
          type: string
          enum: [pending, verified, failed]
        verification_records:
          type: array
          items:
            $ref: "#/components/schemas/DnsRecord"
          example:
            - type: TXT
              name: "@"
              value: "v=spf1 include:amazonses.com ~all"
              purpose: spf
            - type: CNAME
              name: "dkim._domainkey"
              value: "dkim.amazonses.com"
              purpose: dkim
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    UsageResponse:
      type: object
      required: [period, emails_sent, sms_sent, sms_credits_balance]
      properties:
        period:
          type: string
        emails_sent:
          type: integer
        sms_sent:
          type: integer
        sms_credits_balance:
          type: integer
        overage_cost_cents:
          type: integer
    LimitsResponse:
      type: object
      required: [plan, max_users, max_clients, retention_days]
      properties:
        plan:
          type: string
          enum: [starter, pro, business]
        max_users:
          type: integer
        max_clients:
          type: integer
        retention_days:
          type: integer
        sms_included:
          type: integer
        sms_addon_enabled:
          type: boolean
        sms_price_cents:
          type: integer
...
    RegisterRequest:
      type: object
      required: [email, password, name, workspace_name]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
        name:
          type: string
        workspace_name:
          type: string
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
    RefreshRequest:
      type: object
      required: [refresh_token]
      properties:
        refresh_token:
          type: string
    LogoutRequest:
      type: object
      properties:
        refresh_token:
          type: string
    OauthExchangeRequest:
      type: object
      required: [code, state]
      properties:
        code:
          type: string
        state:
          type: string
    PasswordForgotRequest:
      type: object
      required: [email]
      properties:
        email:
          type: string
          format: email
    PasswordResetRequest:
      type: object
      required: [token, new_password]
      properties:
        token:
          type: string
        new_password:
          type: string
          format: password
...
    SendingDomainCreateRequest:
      type: object
      required: [domain]
      properties:
        domain:
          type: string
...
paths:
  /auth/register:
    post:
      tags: [Auth]
      summary: Register user and workspace
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
            example:
              email: owner@example.com
              password: StrongPass123
              name: Owner User
              workspace_name: Acme Workspace
      responses:
        "200":
          description: Auth response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          $ref: "#/components/responses/BadRequestResponse"
        "409":
          $ref: "#/components/responses/ConflictResponse"
  /auth/login:
    post:
      tags: [Auth]
      summary: Login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
            example:
              email: owner@example.com
              password: StrongPass123
      responses:
        "200":
          description: Auth response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          $ref: "#/components/responses/BadRequestResponse"
        "401":
          $ref: "#/components/responses/UnauthorizedResponse"
  /auth/refresh:
    post:
      tags: [Auth]
      summary: Refresh tokens
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshRequest"
            example:
              refresh_token: 6f7c0b40-1f32-4f1b-8ab1-89b1b7cc0b91
      responses:
        "200":
          description: Auth response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          $ref: "#/components/responses/BadRequestResponse"
        "401":
          $ref: "#/components/responses/UnauthorizedResponse"
  /auth/logout:
    post:
      tags: [Auth]
      summary: Logout (revoke refresh token)
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LogoutRequest"
            example:
              refresh_token: 6f7c0b40-1f32-4f1b-8ab1-89b1b7cc0b91
      responses:
        "200":
          description: Status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatusResponse"
        "400":
          $ref: "#/components/responses/BadRequestResponse"
        "401":
          $ref: "#/components/responses/UnauthorizedResponse"
  /auth/me:
    get:
      tags: [Auth]
      summary: Current user and workspace context
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
      responses:
        "200":
          description: Me response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MeResponse"
        "401":
          $ref: "#/components/responses/UnauthorizedResponse"
...
  /auth/oauth/google/exchange:
    post:
      tags: [Auth]
      summary: Exchange OAuth code
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OauthExchangeRequest"
      responses:
        "200":
          description: Auth response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          $ref: "#/components/responses/BadRequestResponse"
  /auth/password/forgot:
    post:
      tags: [Auth]
      summary: Request password reset
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PasswordForgotRequest"
            example:
              email: owner@example.com
      responses:
        "200":
          description: Status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatusResponse"
        "400":
          $ref: "#/components/responses/BadRequestResponse"
  /auth/password/reset:
    post:
      tags: [Auth]
      summary: Reset password
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PasswordResetRequest"
            example:
              token: reset-token
              new_password: NewStrongPass123
      responses:
        "200":
          description: Status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatusResponse"
        "400":
          $ref: "#/components/responses/BadRequestResponse"
...
  /sending-domains:
    get:
      tags: [SendingDomains]
      summary: List sending domains
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
      responses:
        "200":
          description: Domain list
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/SendingDomainResponse"
        "401":
          $ref: "#/components/responses/UnauthorizedResponse"
        "403":
          $ref: "#/components/responses/ForbiddenResponse"
    post:
      tags: [SendingDomains]
      summary: Add sending domain
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendingDomainCreateRequest"
            example:
              domain: example.com
      responses:
        "200":
          description: Domain response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SendingDomainResponse"
        "400":
          $ref: "#/components/responses/BadRequestResponse"
        "401":
          $ref: "#/components/responses/UnauthorizedResponse"
        "403":
          $ref: "#/components/responses/ForbiddenResponse"
        "409":
          $ref: "#/components/responses/ConflictResponse"
  /sending-domains/{domainId}:
    get:
      tags: [SendingDomains]
      summary: Get sending domain
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
        - name: domainId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Domain response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SendingDomainResponse"
        "401":
          $ref: "#/components/responses/UnauthorizedResponse"
        "403":
          $ref: "#/components/responses/ForbiddenResponse"
  /sending-domains/{domainId}/verify:
    post:
      tags: [SendingDomains]
      summary: Check domain verification
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
        - name: domainId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Domain response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SendingDomainResponse"
        "401":
          $ref: "#/components/responses/UnauthorizedResponse"
        "403":
          $ref: "#/components/responses/ForbiddenResponse"
...