# src/main/java/com/sgagestudio/warm_follow_backend/dto/DnsRecord.java
package com.sgagestudio.warm_follow_backend.dto;

public record DnsRecord(
        DnsRecordType type,
        String name,
        String value,
        Integer ttl,
        Integer priority,
        DnsRecordPurpose purpose
) {
}